added backup delete action

+++ src/AZMongoBak/ClientApp/src/pages/database-page/tabs/backup-tab/backup-details.tsx
@@ -1,8 +1,14 @@
 import { Box, Button, Text } from "@primer/react";
-import { Backup } from "../../../../data-models/db-models";
+import { Backup, BackupInfo } from "../../../../data-models/db-models";
 import { TrashIcon, XIcon } from "@primer/octicons-react";
+import { useState } from "react";
+import { InfoMessage } from "../../../../shared/messages/info-message";
+import { ErrorMessage } from "../../../../shared/messages/error-message";
+import { APIService } from "../../../../auth/api-service";
+import { useMsal } from "@azure/msal-react";
 
 type TBackupDetailsProps = {
+    backup_info?: BackupInfo;
     backup: Backup;
     onClose: () => void;
     onDelete: () => void;
@@ -10,11 +16,28 @@ type TBackupDetailsProps = {
 
 export const BackupDetails = (props: TBackupDetailsProps) => {
 
-    const { backup, onClose, onDelete } = props;
+    const { backup, backup_info, onClose, onDelete } = props;
+    const { instance, accounts } = useMsal();
+    
+    const [btn_disabled, setBtnDisabled] = useState(false);
+    const [info_msg, setInfo] = useState<string>();
+    const [error_msg, setError] = useState<string>();
 
     const HandleOnDelete = async () => {
-        // Placeholder TODO
-        onDelete();
+        if (backup_info) {
+            setBtnDisabled(true);
+            setError(undefined);
+            setInfo("Deleting backup, please wait ..");
+
+            if (await backup.DeleteAsync(backup_info.oid, new APIService(instance, accounts))) {
+                onDelete();
+            }
+            else {
+                setBtnDisabled(false);
+                setInfo(undefined);
+                setError("Failed to delete this backup");
+            }
+        }
     };
 
     return (
@@ -77,13 +100,26 @@ export const BackupDetails = (props: TBackupDetailsProps) => {
                         }
                     </Box>
 
-                    <Button
-                        onClick={HandleOnDelete}
-                        sx={{mt: 4}}
-                        leadingIcon={() => <TrashIcon />}
-                        variant="danger">
-                        Delete backup
-                    </Button>
+                    { backup.job.done &&
+                        <>
+                            <Button
+                                disabled={btn_disabled}
+                                onClick={HandleOnDelete}
+                                sx={{mt: 4}}
+                                leadingIcon={() => <TrashIcon />}
+                                variant="danger">
+                                Delete backup
+                            </Button>
+
+                            <InfoMessage 
+                                style={{ marginTop: "20px" }}
+                                message={info_msg} />
+
+                            <ErrorMessage 
+                                style={{ marginTop: "20px" }}
+                                message={error_msg} />
+                        </>
+                    }
                 </div>
             }
         </>
